refactor(langchain): dedupe doc-conversion helpers into _common
The three result-to-Document helpers (_payload_to_doc, _results_to_docs,
_results_to_docs_with_score) were byte-identical clones in both
multi_query_ops.py and search_ops.py. Move them to the package's _common
module (single definition) and import in both, eliminating the duplication
that kept multi_query_ops.py at grade D. No behaviour change; langchain
integration suite unchanged (195 pass, 7 pre-existing env failures).
