+++ integrations/langchain/src/langchain_velesdb/search_ops.py
@@ -14,7 +14,11 @@ from typing import Any, List, Optional, Tuple
 
 from langchain_core.documents import Document
 
-from langchain_velesdb._common import payload_to_doc_parts
+from langchain_velesdb._common import (
+    payload_to_doc_parts,
+    _results_to_docs,
+    _results_to_docs_with_score,
+)
 from langchain_velesdb.multi_query_ops import MultiQueryOpsMixin
 from langchain_velesdb.security import (
     validate_k,
@@ -40,22 +44,6 @@ except (ImportError, AttributeError):  # pragma: no cover — optional dependenc
     _SearchOptions = None  # type: ignore[assignment]
 
 logger = logging.getLogger(__name__)
-
-
-def _payload_to_doc(result: dict) -> Document:
-    """Convert a single search result dict to a LangChain Document."""
-    text, metadata = payload_to_doc_parts(result)
-    return Document(page_content=text, metadata=metadata)
-
-
-def _results_to_docs(results: List[dict]) -> List[Document]:
-    """Convert a list of search result dicts to Documents."""
-    return [_payload_to_doc(r) for r in results]
-
-
-def _results_to_docs_with_score(results: List[dict]) -> List[Tuple[Document, float]]:
-    """Convert a list of search result dicts to (Document, score) tuples."""
-    return [(_payload_to_doc(r), r.get("score", 0.0)) for r in results]
 
 
 class SearchOpsMixin(MultiQueryOpsMixin):
